mcp-openai-gemini/pyproject.toml
[project]
name = "mcp-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.400",
    "mcp[cli]>=1.6.0",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
]
